pyproject.toml
[project]
name = "aibrowseagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "browser-use>=0.9.7",
    "certifi>=2025.11.12",
    "django>=5.2.8",
    "google-api-python-client>=2.187.0",
    "google-auth-httplib2>=0.2.1",
    "google-auth-oauthlib>=1.2.2",
    "opencv-python>=4.12.0.88",
    "pillow>=12.0.0",
    "playwright-stealth==1.0.5",
    "pymupdf>=1.26.6",
    "setuptools>=80.9.0",
]
